# docs/api/openapi.v2.json
{
  "components": {
    "schemas": {
...
      "AdminPeopleListResponseV2": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "limit": {
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/AdminPersonSummaryV2"
            },
            "title": "People",
            "type": "array"
          },
          "total_count": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count"
          }
        },
        "required": [
          "people",
          "limit",
          "offset",
          "count",
          "has_more"
        ],
        "title": "AdminPeopleListResponseV2",
        "type": "object"
      },
      "AdminPeopleReadProblemDetailV2": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retry_after_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Ms"
          },
          "retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retryable"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "status",
          "message",
          "trace_id",
          "request_id"
        ],
        "title": "AdminPeopleReadProblemDetailV2",
        "type": "object"
      },
      "AdminPeopleReadProblemResponseV2": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/AdminPeopleReadProblemDetailV2"
          }
        },
        "required": [
          "detail"
        ],
        "title": "AdminPeopleReadProblemResponseV2",
        "type": "object"
      },
...
      "AdminPersonResponseV2": {
        "additionalProperties": false,
        "properties": {
          "person": {
            "$ref": "#/components/schemas/AdminPersonV2"
          }
        },
        "required": [
          "person"
        ],
        "title": "AdminPersonResponseV2",
        "type": "object"
      },
      "AdminPersonSummaryV2": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "external_ids": {
            "additionalProperties": true,
            "title": "External Ids",
            "type": "object"
          },
          "full_name": {
            "minLength": 1,
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "known_for": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Known For"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "full_name",
          "created_at",
          "updated_at"
        ],
        "title": "AdminPersonSummaryV2",
        "type": "object"
      },
      "AdminPersonV2": {
        "additionalProperties": false,
        "properties": {
          "alternative_names": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alternative Names"
          },
          "biography": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Biography"
          },
          "birthday": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthday"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "external_ids": {
            "additionalProperties": true,
            "title": "External Ids",
            "type": "object"
          },
          "full_name": {
            "minLength": 1,
            "title": "Full Name",
            "type": "string"
          },
          "gender": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "homepage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Homepage"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "known_for": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Known For"
          },
          "place_of_birth": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Of Birth"
          },
          "profile_image_url": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Image Url"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "full_name",
          "created_at",
          "updated_at"
        ],
        "title": "AdminPersonV2",
        "type": "object"
      },
...
      "IdentityMatchKind": {
        "enum": [
          "canonical",
          "alias"
        ],
        "title": "IdentityMatchKind",
        "type": "string"
      },
...
      "IdentityProblemResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/IdentityProblemDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "IdentityProblemResponse",
        "type": "object"
      },
...
      "PersonIdentityResponse": {
        "additionalProperties": false,
        "properties": {
          "canonical_path": {
            "title": "Canonical Path",
            "type": "string"
          },
          "canonical_slug": {
            "title": "Canonical Slug",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "match_kind": {
            "$ref": "#/components/schemas/IdentityMatchKind"
          },
          "person_id": {
            "format": "uuid",
            "title": "Person Id",
            "type": "string"
          },
          "requested_slug": {
            "title": "Requested Slug",
            "type": "string"
          },
          "resource_type": {
            "const": "person",
            "title": "Resource Type",
            "type": "string"
          },
          "show_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonShowContext"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "resource_type",
          "person_id",
          "full_name",
          "requested_slug",
          "canonical_slug",
          "match_kind",
          "canonical_path"
        ],
        "title": "PersonIdentityResponse",
        "type": "object"
      },
...
      "PersonShowContext": {
        "additionalProperties": false,
        "properties": {
          "canonical_slug": {
            "title": "Canonical Slug",
            "type": "string"
          },
          "show_id": {
            "format": "uuid",
            "title": "Show Id",
            "type": "string"
          },
          "show_name": {
            "title": "Show Name",
            "type": "string"
          }
        },
        "required": [
          "show_id",
          "show_name",
          "canonical_slug"
        ],
        "title": "PersonShowContext",
        "type": "object"
      },
...
      "RecentPeopleListResponseV2": {
        "additionalProperties": false,
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/RecentPeoplePaginationV2"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/RecentPersonV2"
            },
            "title": "People",
            "type": "array"
          }
        },
        "required": [
          "people",
          "pagination"
        ],
        "title": "RecentPeopleListResponseV2",
        "type": "object"
      },
      "RecentPeoplePaginationV2": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          },
          "limit": {
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          }
        },
        "required": [
          "limit",
          "count"
        ],
        "title": "RecentPeoplePaginationV2",
        "type": "object"
      },
      "RecentPeopleProblemDetailV2": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retry_after_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Ms"
          },
          "retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retryable"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "status",
          "message",
          "trace_id",
          "request_id"
        ],
        "title": "RecentPeopleProblemDetailV2",
        "type": "object"
      },
      "RecentPeopleProblemResponseV2": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/RecentPeopleProblemDetailV2"
          }
        },
        "required": [
          "detail"
        ],
        "title": "RecentPeopleProblemResponseV2",
        "type": "object"
      },
      "RecentPersonV2": {
        "additionalProperties": false,
        "properties": {
          "first_viewed_at": {
            "title": "First Viewed At",
            "type": "string"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "known_for": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Known For"
          },
          "last_viewed_at": {
            "title": "Last Viewed At",
            "type": "string"
          },
          "person_id": {
            "format": "uuid",
            "title": "Person Id",
            "type": "string"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "show_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Context"
          },
          "view_count": {
            "minimum": 0.0,
            "title": "View Count",
            "type": "integer"
          }
        },
        "required": [
          "person_id",
          "full_name",
          "known_for",
          "photo_url",
          "show_context",
          "view_count",
          "first_viewed_at",
          "last_viewed_at"
        ],
        "title": "RecentPersonV2",
        "type": "object"
      },
      "RecordRecentPersonResponseV2": {
        "additionalProperties": false,
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "RecordRecentPersonResponseV2",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "InternalAdminBearer": {
        "bearerFormat": "JWT",
        "description": "An allowlisted Supabase admin JWT or a TRR-signed internal-admin JWT. The backend applies the same InternalAdminUser authorization policy to both.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend API for The Reality Report - reality TV data and surveys",
    "title": "The Reality Report API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/admin/people": {
      "get": {
        "operationId": "listAdminPeopleV2",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 500,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPeopleListResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPeopleReadProblemResponseV2"
                }
              }
            },
            "description": "The people read request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPeopleReadProblemResponseV2"
                }
              }
            },
            "description": "The people read request failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPeopleReadProblemResponseV2"
                }
              }
            },
            "description": "The people store is unavailable."
          }
        },
        "security": [
          {
            "InternalAdminBearer": []
          }
        ],
        "summary": "List or prefix-search core people",
        "tags": [
          "admin-core-people-reads-v2"
        ]
      }
    },
...
    "/api/v2/admin/people/{person_id}": {
      "get": {
        "operationId": "getAdminPersonV2",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPersonResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPeopleReadProblemResponseV2"
                }
              }
            },
            "description": "The people read request is invalid."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPeopleReadProblemResponseV2"
                }
              }
            },
            "description": "The requested person was not found."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPeopleReadProblemResponseV2"
                }
              }
            },
            "description": "The people read request failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPeopleReadProblemResponseV2"
                }
              }
            },
            "description": "The people store is unavailable."
          }
        },
        "security": [
          {
            "InternalAdminBearer": []
          }
        ],
        "summary": "Get a core person",
        "tags": [
          "admin-core-people-reads-v2"
        ]
      }
    },
...
    "/api/v2/admin/recent-people": {
      "get": {
        "operationId": "listAdminRecentPeopleV2",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentPeopleListResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentPeopleProblemResponseV2"
                }
              }
            },
            "description": "The recent-people request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentPeopleProblemResponseV2"
                }
              }
            },
            "description": "The recent-people request failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentPeopleProblemResponseV2"
                }
              }
            },
            "description": "The recent-people store is unavailable."
          }
        },
        "security": [
          {
            "InternalAdminBearer": []
          }
        ],
        "summary": "List Recent People",
        "tags": [
          "admin-recent-people-v2"
        ]
      },
      "post": {
        "operationId": "recordAdminRecentPersonV2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "personId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "showId": {
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                },
                "required": [
                  "personId"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordRecentPersonResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentPeopleProblemResponseV2"
                }
              }
            },
            "description": "The recent-people request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentPeopleProblemResponseV2"
                }
              }
            },
            "description": "The recent-people request failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentPeopleProblemResponseV2"
                }
              }
            },
            "description": "The recent-people store is unavailable."
          }
        },
        "security": [
          {
            "InternalAdminBearer": []
          }
        ],
        "summary": "Record Recent Person",
        "tags": [
          "admin-recent-people-v2"
        ]
      }
    },
...
    "/api/v2/identities/people/{slug}": {
      "get": {
        "description": "Resolves a canonical slug or direct alias. Optional show_id or show_slug context narrows collisions to people linked to that show's cast. The endpoint rejects requests that provide both contexts.",
        "operationId": "resolvePublicPersonIdentityV2",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "example": "brandi-glanville",
              "maxLength": 160,
              "type": "string"
            }
          },
          {
            "description": "Optional show UUID used to narrow person collisions.",
            "in": "query",
            "name": "show_id",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Optional show slug or direct alias used to narrow person collisions.",
            "in": "query",
            "name": "show_slug",
            "required": false,
            "schema": {
              "maxLength": 160,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonIdentityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityProblemResponse"
                }
              }
            },
            "description": "Invalid slug or conflicting show context."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityProblemResponse"
                }
              }
            },
            "description": "No public identity matched the request."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityProblemResponse"
                }
              }
            },
            "description": "A direct alias matched multiple identities."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityProblemResponse"
                }
              }
            },
            "description": "The resolver could not complete the request."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityProblemResponse"
                }
              }
            },
            "description": "The identity store is temporarily unavailable."
          }
        },
        "summary": "Resolve a public person identity",
        "tags": [
          "public-identities-v2"
        ]
      }
    },
...
  }
}